Decompose/SplitModule: include modID in partioning, change modID's.
ModID's now contain 'path' they took for splitting,
which I don't have any use for but might be interesting.

(OrigModID)_x_y_z

Means OrigModID was split and this was Module 'x' (0-index) of
the first split, 'y' of the next, and so on.

Recomputing MD5 of module hash is wasteful but probably not
a significant concern.